requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
pathspec==0.10.3
platformdirs==2.6.0
psycopg2==2.9.5
pusher==3.3.0
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.1.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
